change how script looks for words, different separator
Finding words didn't work for me, result was always empty, so no records
were added to database, and apkg were created without definition. To
overcome it I've changed faulty jquery (I suppose) to separating words
with substring, which searches for closing tag of h2 (opening tag was
the most suitable for separator for my dictionary, as it was before
every word, after the word comes the closing tag), and then assigning
substring before closing tag to word, and the rest after the closing tag
to the definition. This way query now works like a charm, but it might
not work for every dictionary, as layout might differ, so further
research is required.